Changed to use 3D graph
